refactor: - 키워드, 카테고리로 검색 ⇒ 로그인 안했을 때도 가능하게(추천, 관심 빼고)           - bigCategoryID로만 검색가능하게도 하기           - classID가 없을 때 관심클래스에 저장 안되도록 하기(예외처리)           - ClickedHobbyClass domain 추가
Former-commit-id: 96ea413b81f3ade3d6f22e7d3938bb35db8cc9a2

+++ Backend/src/main/java/com/ssafy/buki/service/InterestService.java
@@ -83,14 +83,15 @@ public class InterestService {
     // 3. Post 관심 클래스 설정
     public void setInterestClass(InterestHobbyClassReqDto interestHobbyClassReq, User user) {
         Long id = interestHobbyClassReq.getHobbyClassId();
+        HobbyClass hobbyClass = hobbyClassRepository.findHobbyClassById(id);
+        if (hobbyClass == null) throw new BusinessException(NOT_RIGHT_DATA);
+
         InterestHobbyClass interestHobbyClass = interestHobbyClassRepository.findInterestHobbyClassByHobbyClassIdAndUserId(id, user.getId());
         if (interestHobbyClassReq.isInterest()) { //관심 클래스일 경우
             if (interestHobbyClass == null) throw new BusinessException(NOT_RIGHT_DATA);
             interestHobbyClassRepository.deleteInterestHobbyClassByHobbyClassIdAndUserId(id, user.getId());
         } else { //관심 클래스가 아닐 경우
             if (interestHobbyClass != null) throw new BusinessException(ALREADY_HAS_DATA);
-            HobbyClass hobbyClass = hobbyClassRepository
-                    .findHobbyClassById(id);
 
             interestHobbyClassRepository.save(InterestHobbyClass.builder()
                     .user(user)
